fix: move code from auth service to jwt service
