+++ iks_projekat_tim_24_uberApp/src/app/components/jwt-service.service.ts
@@ -1,16 +1,40 @@
 import { Injectable } from '@angular/core';
 import jwtDecode from 'jwt-decode';
 import { User } from '../user';
+import { Subject } from 'rxjs';
 
 @Injectable({
   providedIn: 'root'
 })
 export class JwtService {
+  private _isLoggedIn = false;
+  private _isLoggedInSubject = new Subject<boolean>();
+  isLoggedIn$ = this._isLoggedInSubject.asObservable();
+
+  get isLoggedIn() {
+    return this._isLoggedIn;
+  }
+
+  set isLoggedIn(value: boolean) {
+    this._isLoggedIn = value;
+    this._isLoggedInSubject.next(value);
+  }
+
   setJwt(jwt: string){
     localStorage.setItem('jwt',jwt);
+    this.updateLoggedIn();
   }
+
   getJwt(){
     return localStorage.getItem('jwt');
+  }
+
+  updateLoggedIn()
+  {
+    let role = this.getRole();
+    if (role === "DRIVER" || role === "ADMIN" || role==="USER") {
+      this.isLoggedIn = true;
+    }
   }
 
   getEmail() : string | null{
