+++ iks_projekat_tim_24_uberApp/src/app/components/login/login.component.ts
@@ -6,7 +6,6 @@ import { UserService } from 'src/app/user.service';
 import { environment } from 'src/environments/environment';
 import { JwtService } from '../jwt-service.service';
 import { loginResponse } from './loginResponse';
-import { AuthService } from 'src/app/services/auth.service';
 
 
 @Component({
@@ -18,7 +17,7 @@ export class LoginComponent implements OnInit {
 
   loginForm: FormGroup;
   
-  constructor(private http: HttpClient, private jwtSetvice: JwtService, private router:Router, private userService:UserService, private authService:AuthService) { 
+  constructor(private http: HttpClient, private jwtService: JwtService, private router:Router, private userService:UserService) { 
   }
 
   ngOnInit(): void {
@@ -34,14 +33,19 @@ export class LoginComponent implements OnInit {
     const password = this.loginForm.get('password')?.value;
     
     const response = await this.http.post(environment.apiBaseUrl+'api/user/login', {email:email, password:password}).toPromise() as loginResponse;
-    this.jwtSetvice.setJwt(response.accessToken);
-    //this.router.navigate(['/']);
+    this.jwtService.setJwt(response.accessToken);
 
-    let role = this.jwtSetvice.getRole();
+    this.routeUsers();
 
-    if (role === "DRIVER" || role === "ADMIN" || role==="USER") {
-      this.authService.isLoggedIn = true;
     }
+    catch (HttpErrorResponse){
+      console.error(HttpErrorResponse);
+    }
+  };
+
+  routeUsers()
+  {
+    let role = this.jwtService.getRole();
 
     if (role === "DRIVER")
     {
@@ -55,12 +59,7 @@ export class LoginComponent implements OnInit {
     {
       this.router.navigate(['/registered-home'])
     }
-
-    }
-    catch (HttpErrorResponse){
-      console.error(HttpErrorResponse);
-    }
-  };
+  }
   
 }
