+++ iks_projekat_tim_24_uberApp/src/app/components/navbar-registered/navbar-registered.component.html
@@ -3,8 +3,9 @@
     <i class="fa fa-solid fa-bell" [ngClass]="{'no-notifications': !hasNotification, 'has-notifications': hasNotification}" [routerLink]="['/notifications']"></i>
     <div class="profile-pic" mat-raised-button  [matMenuTriggerFor]="menu"></div>
 </div>
-<mat-menu #menu="matMenu">
+<mat-menu #menu="matMenu" >
     <button mat-menu-item [routerLink]="['/notifications']">jozsefe</button>
     <button mat-menu-item>vlaf</button>
     <button mat-menu-item>mjehao</button>
 </mat-menu>
+
